Add timer test hacks
I think this testing was valuable, even though it didn't yet yield the results I was looking for.

I think the next thing to try will be to extract all of the sample's
data from the ROM, along with the timer register settings in the code,
and see if we can get the sample playing back in a .wav file at the
correct rate to work out what the unknowns here might be. This should
point us in the right direction wrt sample rate, timer frequency, etc.